+++ lib/assets_handlers/custom_icons.dart
@@ -29,4 +29,8 @@ class CustomIcons {
       IconData(0xe801, fontFamily: _kFontFam, fontPackage: _kFontPkg);
   static const IconData swords =
       IconData(0xe802, fontFamily: _kFontFam, fontPackage: _kFontPkg);
+  static const IconData capital =
+      IconData(0xe803, fontFamily: _kFontFam, fontPackage: _kFontPkg);
+  static const IconData crown =
+      IconData(0xe804, fontFamily: _kFontFam, fontPackage: _kFontPkg);
 }