+++ lib/pages/general/generators.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:realm_roller/custom_widgets/route_builder/route_builder.dart';
 import 'package:realm_roller/pages/deities/deity_generation/deity_generation_page.dart';
+import 'package:realm_roller/pages/emblem/emblem_generation/emblem_generation_page.dart';
 import 'package:realm_roller/pages/guilds/guild_generation/guild_generation_page.dart';
+import 'package:realm_roller/pages/kingdoms/kingdom_generation/kingdom_generation_page.dart';
 import 'package:realm_roller/pages/landscapes/landscape_generation/landscape_generation_page.dart';
 import 'package:realm_roller/pages/locations/locations_generation/location_generation_page.dart';
 import 'package:realm_roller/pages/names/names_generation/npc_names_generation_page.dart';
@@ -25,6 +27,8 @@ class GeneratorsPage extends StatelessWidget {
       "Landscapes": LandscapeGenerationPage(),
       "Deities": DeityGenerationPage(),
       "Guilds": GuildGenerationPage(),
+      "Kingdoms": KingdomGenerationPage(),
+      "Emblems": EmblemGenerationPage(),
     };
     final titles = generators.keys.toList();