+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog 
 
+## 0.9.0
+
+- Added kingdom generation
+
+- Added emblem generation
+
 ## 0.8.0
 
 - Added guild generation